+++ packages/orq-rc/RELEASES.md
@@ -7739,3 +7739,13 @@ Based on:
 - [python v4.9.0-rc.41] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.41] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.41 - packages/orq-rc
+
+## 2026-05-11 13:47:39
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.42] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.42] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.42 - packages/orq-rc